pyproject.toml
[tool.poetry]
name = "tira-labra"
version = "0.1.0"
description = ""
authors = ["Aarni <dabeastelele@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pygame = "^2.3.0"
invoke = "^2.0.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
coverage = "^7.2.2"
pylint = "^2.17.1"
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.autopep8]
max_line_length = 80
in-place = true
recursive = true
aggressive = 3
